fix(validation): Restructure validation file structure
