+++ server/controllers/users.js
@@ -5,7 +5,7 @@ import jwt from 'jsonwebtoken';
 
 import EmailVerificationController from './emailVerificationController';
 import isValidNumber from '../utils/is_valid_number';
-import UserValidation from '../validation/users';
+import UserValidation from '../utils/validation';
 import trimInput from '../utils/trim_input';
 import TokenHelper from '../utils/TokenHelper';
 import { User } from '../models';
